distutils: customize compiler with syslog-ng specific dirs
We are trying to build `pyeditline`, which does not support placing
headers and libraries in non-standard locations. We can "inject" our
directories to `distutils`, which is enough in this case.

In detail:
`Configure()` is called multiple times in pyeditline's source:
https://github.com/mark-nicholson/python-editline/blob/master/setupext/autoconf.py#L279

Although a `compiler` is seemingly passed to `Configure()`, it is `None`
in reality.

`Configure()` is from another repo. If its constructor does not get a
valid `compiler`, it creates a new, and calls `customize_compiler()`:
https://github.com/mark-nicholson/python-hostconf/blob/master/hostconf/configure.py#L82

That is why we can put our custom directories there.

I have tried to add the directories with CFLAGS, LDFLAGS, etc... it did
not work, unfortunately.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>