distutils: remove --coverage cflag
When we apply this flag, all the objects built need to link against -lgcov.

There are symbol checks in python-hostconf/configure.py
(used for pyeditline), which try to build and link with the bare minimum
linking options, to auto-detect the available and necessary linking flags.
These checks always fail, because the binaries are missing the -lgcov flag.

When we are checking the coverage, we are not interested in pyeditline's,
thus we can simply disable it.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>